cheri: Apply the freepage.cocci semantic patch
Calls to free_page and free_pages that cast its (pointer)
argument to unsigned long must cast it to uintptr_t instead.

Apply like this:

	make COCCI=scripts/coccinelle/cheri/free_page.cocci \
		MODE=patch coccicheck

+++ arch/arm64/mm/mmu.c
@@ -842,7 +842,7 @@ static void free_hotplug_page_range(struct page *page, size_t size,
 		vmem_altmap_free(altmap, size >> PAGE_SHIFT);
 	} else {
 		WARN_ON(PageReserved(page));
-		free_pages((unsigned long)page_address(page), get_order(size));
+		free_pages((uintptr_t)page_address(page), get_order(size));
 	}
 }
