+++ arch/x86/mm/pat/set_memory.c
@@ -1414,7 +1414,7 @@ static bool try_to_free_pte_page(pte_t *pte)
 		if (!pte_none(pte[i]))
 			return false;
 
-	free_page((unsigned long)pte);
+	free_page((uintptr_t)pte);
 	return true;
 }
 
@@ -1426,7 +1426,7 @@ static bool try_to_free_pmd_page(pmd_t *pmd)
 		if (!pmd_none(pmd[i]))
 			return false;
 
-	free_page((unsigned long)pmd);
+	free_page((uintptr_t)pmd);
 	return true;
 }
