+++ fs/ocfs2/dlm/dlmmaster.c
@@ -2738,7 +2738,7 @@ leave:
 		wake_up(&res->wq);
 
 	if (mres)
-		free_page((unsigned long)mres);
+		free_page((uintptr_t)mres);
 
 	dlm_put(dlm);
